agregado libreria

+++ js/main.js
@@ -8,8 +8,11 @@ class Producto{
         this.img = imgProdu;   
         this.cantidad = 0; 
     }
-    manejarcantidad(comp){
+    manejarCantidad(comp){
         this.cantidad=comp;
+    }
+    manejarStock(cant,vaciar){
+        vaciar ? this.stock += parseInt(cant) : this.stock -= cant
     }
 }
 
@@ -33,17 +36,26 @@ class Totales{
 let precioTotal = 0;
 let descuento = false;
 let precioTotal2 = 0;
+let vaciar = false;
+
+let listaProductos = [];
+listaCompraDos = [];
+produc();
+recargar = localStorage.getItem("Recargue")
 
 
-let producto1 = new Producto(1,"conjunto deportivo",7199,30,"hombre","./fotos/conjuntodeportivo.jpg");
-let producto2 = new Producto(2,"saco",6000,10,"hombre","./fotos/Saco.jpg");
-let producto3 = new Producto(3,"remera hombre",800,100,"hombre","./fotos/Remera.jpg");
-let producto4 = new Producto(4,"jean",2000,26,"hombre","./fotos/jean-hombre.jpg");
-let producto5 = new Producto(5,"remera deportiva mujer",1050,45,"mujer","./fotos/remeradeportiva.jpg");
-let producto6 = new Producto(6,"remera deportiva",1300,45,"hombre","./fotos/remeradeporte.jpg");
-let producto7 = new Producto(7,"jean mujer",1800,20,"mujer","./fotos/jean.jpg");
+function produc(){
+    let producto1 = new Producto(1,"conjunto deportivo",7199,30,"hombre","./fotos/conjuntodeportivo.jpg");
+    let producto2 = new Producto(2,"saco",6000,10,"hombre","./fotos/Saco.jpg");
+    let producto3 = new Producto(3,"remera hombre",800,100,"hombre","./fotos/Remera.jpg");
+    let producto4 = new Producto(4,"jean",2000,26,"hombre","./fotos/jean-hombre.jpg");
+    let producto5 = new Producto(5,"remera deportiva mujer",1050,45,"mujer","./fotos/remeradeportiva.jpg");
+    let producto6 = new Producto(6,"remera deportiva",1300,45,"hombre","./fotos/remeradeporte.jpg");
+    let producto7 = new Producto(7,"jean mujer",1800,20,"mujer","./fotos/jean.jpg");
+    listaProductos = [producto1,producto2,producto3,producto4,producto5,producto6,producto7];
+}
 
-let listaProductos = [producto1,producto2,producto3,producto4,producto5,producto6,producto7];
+//let listaProductos = [producto1,producto2,producto3,producto4,producto5,producto6,producto7];
 let listaCompra = [];
 
 let hombre = document.querySelector('.categoria1');
@@ -84,18 +96,26 @@ function filtroProductos(categoria){
 
 
 function mostrar(listaProducto){
-    for (const producto of listaProducto) {
-        let contenedor = document.createElement("div");
-        contenedor.innerHTML = `
-                                <img src=${producto.img} width="200px" alt="Conjunto_Deportivo" class="imagen"/>
-                                <h3>${producto.nombre}</h3>
-                                <h3>$ ${producto.precio}</h3>
-                                <form class ="formulario">
-                                    <input type="number"name="cantidad" style="width: 4em" min=1 max=${producto.stock} value=0 class ="cantidad1"></input>
-                                    <input type="submit" value="Agregar a carrito" class ="btn1" id=${producto.codigo}></input>
-                                </form>
-                                `
-        vestir.appendChild(contenedor);
+    if(recargar){
+        listaCompraDos = JSON.parse(localStorage.getItem("Carrito"))
+    }
+        for (const producto of listaProducto) {
+            let contenedor = document.createElement("div");
+            contenedor.innerHTML = `
+                                    <img src=${producto.img} width="200px" alt="Conjunto_Deportivo" class="imagen"/>
+                                    <h3>${producto.nombre}</h3>
+                                    <h3>$ ${producto.precio}</h3>
+                                    <form class ="formulario">
+                                        <input type="number"name="cantidad" style="width: 4em" min=1 max=${producto.stock} value=0 class ="cantidad1"></input>
+                                        <input type="submit" value="Agregar a carrito" class ="btn1" id=${producto.codigo}></input>
+                                    </form>
+                                    `
+            vestir.appendChild(contenedor);
+            for (const producto2 of listaCompraDos){
+                if(producto2.codigo == producto.codigo){
+                    producto.manejarStock(producto2.cantidad);
+                }
+            }
     }
     for (const formulario of formularios) {
         formulario.addEventListener("submit",(e)=>e.preventDefault())    
@@ -112,10 +132,21 @@ function validarProducto (e){
     let cod = listaProductos.find(producto=>producto.codigo == id)
     let cantidadPro = e.target.previousSibling.previousSibling.value;
     if(cantidadPro >0){
-        cod.manejarcantidad(cantidadPro)
+        cod.manejarCantidad(cantidadPro)
         if(cod.cantidad>cod.stock){
-            cod.cantidad= cod.stock;
+            Swal.fire({
+                title: 'Error!',
+                text: 'Cantidad pedida es mayor al stock existente',
+                icon: 'info',
+                confirmButtonText: 'Cool'
+              })
+            cod.cantidad = cod.stock;
+            cod.manejarStock(cod.stock);    
         }
+        else{
+            cod.manejarStock(cod.cantidad);
+        }
+        console.log(cod.stock)
         let auxiliar = new Carrito(cod.codigo,cod.nombre,(cod.precio*cod.cantidad),cod.cantidad);
         listaCompra.push(auxiliar);
         
@@ -124,6 +155,7 @@ function validarProducto (e){
         carritoCompra();
         localStorage.setItem("Carrito",JSON.stringify(listaCompra));
         localStorage.setItem("Total",precioTotal);
+        localStorage.setItem("Recargue",true);
     }
     
 }
@@ -132,7 +164,6 @@ function total(){
     if (precioTotal>=100000){
         precioTotal2 = precioTotal * 0.85;
         descuento = true
-        return precioTotal2
     }
 }
 
@@ -140,7 +171,8 @@ function carritoCompra(){
     carritos.innerHTML = ' '
     descuento = false;    
     precioTotal2 = precioTotal;
-    validar = total();
+    //validar = 
+    total();
     let contenedorcarrito1 = document.createElement("div");
     let contenedorcarrito2 = document.createElement("div");
     contenedorcarrito1.innerHTML =`<h3>--------Carrito--------</h3>`
@@ -148,7 +180,7 @@ function carritoCompra(){
     for (const producto of listaCompra) {    
         let contenedorcarrito3 = document.createElement("div");
         contenedorcarrito3.innerHTML = `
-                                <p>${producto.nombre}                  ${producto.cantidad}     $${producto.precio} <form class ="formularioX"> <input type="submit" value="X" class ="btn3 btn" id=${producto.bor}></input></form> </p>
+                                <p>${producto.nombre}   ${producto.cantidad}     $${producto.precio} <form class ="formularioX"> <input type="submit" value="X" class ="btn3 btn" id=${producto.bor}></input></form> </p>
                                 `                        
         carritos.appendChild(contenedorcarrito3);   
     }
@@ -175,21 +207,49 @@ function borrarItem(e){
     localStorage.removeItem("Carrito",JSON.stringify(listaCompra));
     let id = e.target.id;
     let cod = listaCompra.find(producto=>producto.bor == id)
+    for (const produ of listaProductos) {
+        if(produ.codigo === cod.codigo){ 
+            produ.manejarStock((-cod.cantidad)); 
+        }   
+    }
     listaCompra = listaCompra.filter((carritoId) => {
         return carritoId !==cod;
     })
     precioTotal -= cod.precio
     localStorage.setItem("Carrito",JSON.stringify(listaCompra));
     localStorage.setItem("Total",precioTotal);
+    
     carritoCompra();
 }
 
 function vaciarCarrito(){
-    carritos.innerHTML = ' '
-    listaCompra = [];
-    precioTotal = 0;
-    carritoCompra();
-    localStorage.clear();
+
+    Swal.fire({
+        title: 'Vaciar carrito?',
+        text: "Esta por borrar todos sus articulos del carrito",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Borrar'
+      }).then((result) => {
+        if (result.isConfirmed) {
+            carritos.innerHTML = ' '
+            listaCompra = [];
+            listaProducto = [];
+            produc();
+            precioTotal = 0;
+            carritoCompra();
+            localStorage.clear();
+            Swal.fire({
+                position: 'top-end',
+                icon: 'success',
+                title: 'Su carrito fue vaciado',
+                showConfirmButton: false,
+                timer: 1500
+              })
+        }
+      })
 }
 
 function loadCarrito(){
@@ -199,5 +259,7 @@ function loadCarrito(){
     }
     carritoCompra()
 }
+
+
 
 loadCarrito();
